+++ the-blue-alliance-ios/View Controllers/Teams/Team/Team@Event/TeamAtEventViewController.swift
@@ -3,9 +3,9 @@ import Foundation
 import CoreData
 import UIKit
 
-class TeamAtEventViewController: ContainerViewController {
+class TeamAtEventViewController: ContainerViewController, ContainerTeamPushable {
 
-    private let teamKey: TeamKey
+    internal var teamKey: TeamKey
     private let event: Event
 
     // Where should we push to, when clicking the navigation bar from the top. Should be the opposite of what view sent us here
@@ -13,11 +13,9 @@ class TeamAtEventViewController: ContainerViewController {
     private let showDetailEvent: Bool
     private let showDetailTeam: Bool
 
-    private let statusService: StatusService
-    private let urlOpener: URLOpener
-    private let myTBA: MyTBA
-
-    private var teamRequest: URLSessionDataTask?
+    internal var statusService: StatusService
+    internal var urlOpener: URLOpener
+    internal var myTBA: MyTBA
 
 
     // MARK: - Init
@@ -67,12 +65,6 @@ class TeamAtEventViewController: ContainerViewController {
         Analytics.logEvent("team_at_event", parameters: ["event": event.key!, "team": teamKey.key!])
     }
 
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-
-        teamRequest?.cancel()
-    }
-
     // MARK: - Private Methods
 
     @objc private func pushEvent() {
@@ -82,48 +74,6 @@ class TeamAtEventViewController: ContainerViewController {
 
     @objc private func pushTeam() {
         _pushTeam(attemptedToLoadTeam: false)
-    }
-    
-    private func _pushTeam(attemptedToLoadTeam: Bool) {
-        guard let team = teamKey.team else {
-            if attemptedToLoadTeam {
-                showErrorAlert(with: "Unable to load team.")
-            } else {
-                let oldRightBarButtonIcon = navigationItem.rightBarButtonItem
-                DispatchQueue.main.async {
-                    self.navigationItem.rightBarButtonItem = UIBarButtonItem.activityIndicatorBarButtonItem()
-                }
-
-                teamRequest = tbaKit.fetchTeam(key: teamKey.key!, completion: { (team, error) in
-                    let context = self.persistentContainer.newBackgroundContext()
-                    context.performChangesAndWait({
-                        if let team = team {
-                            Team.insert(team, in: context)
-                        }
-                    }, saved: {
-                        // Switch back to our main thread
-                        DispatchQueue.main.async {
-                            self._pushTeam(attemptedToLoadTeam: true) // Try again - but don't get in to a cycle if the request is failing
-                        }
-                    })
-
-                    self.teamRequest = nil
-
-                    // Reset our nav bar item
-                    DispatchQueue.main.async {
-                        self.navigationItem.rightBarButtonItem = oldRightBarButtonIcon
-                    }
-                })
-            }
-            return
-        }
-
-        _pushTeam(team: team)
-    }
-
-    private func _pushTeam(team: Team) {
-        let eventViewController = TeamViewController(team: team, statusService: statusService, urlOpener: urlOpener, myTBA: myTBA, persistentContainer: persistentContainer, tbaKit: tbaKit, userDefaults: userDefaults)
-        navigationController?.pushViewController(eventViewController, animated: true)
     }
 
 }
